requirements.txt
pygame-ce>=2.5.0
PyOpenGL>=3.1.6
psutil>=5.9.0
pillow>=10.0.0
pyinstaller>=6.0.0
pywin32>=306; sys_platform == 'win32'
